Refactor some code and fix issue when no images found
