+++ frontend/src/routes/Infrastructure/Clusters/ManagedClusters/components/ManageNodePoolNodesModal.tsx
